# contract/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Agent hub /v1 contract — single source of truth for hub<->consumer surface",
    "version": "\"2026-09-16\""
  },
  "paths": {
...
    "/v1/skills/{id}/files/{file...}": {
      "put": {
        "description": "Write one file of a skill, creating directories as needed. The hub is a courier: it stores bytes, it does not interpret them.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "file...",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "content": {
                    "type": "string"
                  }
                },
                "required": [
                  "content"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "skillId": {
                      "type": "string"
                    },
                    "path": {
                      "type": "string"
                    },
                    "bytes": {
                      "type": "integer"
                    }
                  },
                  "required": [
                    "skillId",
                    "path",
                    "bytes"
                  ]
                }
              }
            }
          }
        }
      },
      "get": {
        "description": "Read one file of a skill, byte for byte. {file...} is the path inside the skill directory and may contain slashes.",
        "responses": {
          "200": {
            "description": "success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "skillId": {
                      "type": "string"
                    },
                    "path": {
                      "type": "string"
                    },
                    "content": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "skillId",
                    "path",
                    "content"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}